Introducing New Repository, Changing base to SQL

+++ CerteecStore/CerteecStore.Application/Carts/ProductInCartDTO.cs
@@ -10,8 +10,8 @@ namespace CerteecStore.Application.Carts
     {
         public double ProductId { get; set; }
         public string Name { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
-        public double TotalPrice => UnitPrice * Quantity;
+        public decimal TotalPrice => UnitPrice * Quantity;
     }
 }
